APIs/agco-ats.com/v1/python-aiohttp-client/tests/test_logs_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.api_models_api_error import APIModelsApiError
from openapi_server.models.api_models_log import APIModelsLog
from openapi_server.models.api_paged_response_api_models_log import APIPagedResponseAPIModelsLog


pytestmark = pytest.mark.asyncio

async def test_logs_get_log(client):
    """Test case for logs_get_log

    Get a log by ID
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/v2/Logs/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_logs_get_logs(client):
    """Test case for logs_get_logs

    Get the API System logs, most recent first.
    """
    params = [('limit', 56),
                    ('offset', 56)]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/v2/Logs',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_logs_post_log(client):
    """Test case for logs_post_log

    Add a Log entry
    """
    params = [('Message', 'message_example')]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='POST',
        path='/api/v2/Logs',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

